vanControl/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asyncio==4.0.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
deprecation==2.1.0
Django==5.2.7
djangorestframework==3.16.1
dnspython==2.8.0
fonttools==4.54.1
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
kiwisolver==1.4.7
msgpack==1.1.2
multidict==6.7.0
numpy==2.1.2
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
propcache==0.4.1
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pymongo==4.15.3
pyOpenSSL==25.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.2
PyYAML==6.0.3
realtime==2.22.2
redis==7.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
StrEnum==0.4.15
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
websocket==0.2.1
websocket-client==1.9.0
websockets==12.0
yarl==1.22.0
zope.event==6.0
zope.interface==8.0.1
